Dashboard redesign

+++ iron/src/app/shared/components/forms/add-stock-form/add-stock-form.component.ts
@@ -82,7 +82,7 @@ export class AddStockFormComponent implements AfterContentChecked {
     return {
       ticker: ticker,
       initTimestamp: purchaseDateObj.valueOf(),
-      initUnits: parseInt(units),
+      initUnits: parseFloat(units),
     };
   }
 }
